Use exec mixin's builder
The logic for unmarshaling, flags, output, etc is now in the exec
mixin. Bump porter and reuse as much as we can.
