+++ pkg/gcloud/execute_test.go
@@ -4,87 +4,11 @@ import (
 	"bytes"
 	"io/ioutil"
 	"os"
-	"path/filepath"
-	"sort"
 	"testing"
 
-	"github.com/deislabs/porter/pkg/context"
-	"github.com/deislabs/porter/pkg/exec/builder"
-
 	"github.com/deislabs/porter/pkg/test"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	yaml "gopkg.in/yaml.v2"
 )
-
-func TestMixin_UnmarshalInstallAction(t *testing.T) {
-	b, err := ioutil.ReadFile("testdata/install-input.yaml")
-	require.NoError(t, err)
-
-	var action Action
-	err = yaml.Unmarshal(b, &action)
-	require.NoError(t, err)
-
-	require.Equal(t, 1, len(action.Steps))
-	step := action.Steps[0]
-
-	assert.Equal(t, "Configure SSH", step.Description)
-	assert.Equal(t, Groups{"compute"}, step.Groups)
-	assert.Equal(t, "config-ssh", step.Command)
-
-	sort.Sort(step.Flags)
-	assert.Equal(t, builder.Flags{
-		builder.NewFlag("ssh-config-file", "./gce-ssh-config"),
-		builder.NewFlag("ssh-key-file", "./gce-ssh-key")}, step.Flags)
-}
-
-func TestMixin_UnmarshalUpgradeAction(t *testing.T) {
-	b, err := ioutil.ReadFile("testdata/upgrade-input.yaml")
-	require.NoError(t, err)
-
-	var action Action
-	err = yaml.Unmarshal(b, &action)
-	require.NoError(t, err)
-
-	require.Equal(t, 1, len(action.Steps))
-	step := action.Steps[0]
-
-	assert.Equal(t, "Tag VM", step.Description)
-	require.Empty(t, step.Outputs)
-
-	assert.Equal(t, Groups{"compute", "instances"}, step.Groups)
-	assert.Equal(t, "update", step.Command)
-
-	assert.Equal(t, []string{"myinst"}, step.Arguments)
-
-	sort.Sort(step.Flags)
-	assert.Equal(t, builder.Flags{
-		builder.NewFlag("update-labels", "color=blue,ready=true")}, step.Flags)
-}
-
-func TestMixin_UnmarshalUninstallAction(t *testing.T) {
-	b, err := ioutil.ReadFile("testdata/uninstall-input.yaml")
-	require.NoError(t, err)
-
-	var action Action
-	err = yaml.Unmarshal(b, &action)
-	require.NoError(t, err)
-
-	require.Equal(t, 1, len(action.Steps))
-	step := action.Steps[0]
-
-	assert.Equal(t, "Deprovision VM", step.Description)
-	require.Empty(t, step.Outputs)
-
-	assert.Equal(t, Groups{"compute", "instances"}, step.Groups)
-	assert.Equal(t, "delete", step.Command)
-
-	assert.Equal(t, []string{"myinst"}, step.Arguments)
-
-	sort.Sort(step.Flags)
-	assert.Equal(t, builder.Flags{
-		builder.NewFlag("delete-disks", "all")}, step.Flags)
-}
 
 func TestMain(m *testing.M) {
 	test.TestMainWithMockedCommandHandlers(m)
@@ -108,6 +32,7 @@ func TestMixin_Execute(t *testing.T) {
 			"gcloud --quiet compute instances delete myinst --delete-disks all --format json"},
 	}
 
+	defer os.Unsetenv(test.ExpectedCommandEnv)
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
 			os.Setenv(test.ExpectedCommandEnv, tc.wantCommand)
@@ -120,34 +45,4 @@ func TestMixin_Execute(t *testing.T) {
 			require.NoError(t, err, "execute failed")
 		})
 	}
-}
-
-func TestOutputs(t *testing.T) {
-	m := NewTestMixin(t)
-
-	step := Step{
-		Outputs: []Output{
-			{Name: "ids", JsonPath: "$[*].id"},
-			{Name: "names", JsonPath: "$[*].name"},
-		},
-	}
-	output, err := ioutil.ReadFile("testdata/install-output.json")
-	require.NoError(t, err, "could not read testdata")
-
-	err = m.processOutputs(step, bytes.NewBuffer(output))
-	require.NoError(t, err, "processOutputs should not return an error")
-
-	f := filepath.Join(context.MixinOutputsDir, "ids")
-	gotOutput, err := m.FileSystem.ReadFile(f)
-	require.NoError(t, err, "could not read output file %s", f)
-
-	wantOutput := `["1085517466897181794"]`
-	assert.Equal(t, wantOutput, string(gotOutput))
-
-	f = filepath.Join(context.MixinOutputsDir, "names")
-	gotOutput, err = m.FileSystem.ReadFile(f)
-	require.NoError(t, err, "could not read output file %s", f)
-
-	wantOutput = `["porter-test"]`
-	assert.Equal(t, wantOutput, string(gotOutput))
 }
